rename api_security_intro folder back to `api`
This was mistakenly renamed before bc I thought this was the folder for the API Security course. But it's actually for the Graylog API standalone course.

It went undetected bc the API Security course doesn't actually use any external scripts in here- everything is done in the Instruqt track setup script.

Signed-off-by: William Trelawny <william.trelawny@graylog.com>